Make std::hash::RandomState not be align(8) to work around TLS issues on 95
At least on Windows 95 RTM, the TLS variables are only aligned to a 4 byte boundary. Unclear if this affects other systems, but NT4 is fine.

This changes the `hashmap_random_keys` function signature, so update the Linux
implementation to use the new signature as well and keep cross-compiling on
Linux working.

Co-authored-by: Jookia <contact@jookia.org>